==> success_classifier/2_create_pfg_df.ipynb <==
import convokit
from convokit import Corpus, download
import numpy as np
import pandas as pd
import nltk
nltk.download('wordnet', quiet=True)
import re

! pip install contractions
import contractions

## Read in the dataset and convert it to a pandas dataframe

corpus = Corpus(download('persuasionforgood-corpus'))

utterances_df = corpus.get_utterances_dataframe()
conversations_df = corpus.get_conversations_dataframe()

### 1. Write a function to get full conversations as a single string

conversation_ids = corpus.get_conversation_ids()
def get_conversation(conversation_id):
    conversation = ""

    utterances = list(utterances_df[utterances_df['conversation_id'] == conversation_id]['text'])
    for utterance in utterances:
        conversation = conversation + " " + utterance
    
    return conversation

### 2. Write a function to see if any amount was donated during a conversation

def get_conversation_donation(conversation_id):
    er_donation = conversations_df[conversations_df.index == conversation_id]['meta.donation_er'].values[0]
    ee_donation = conversations_df[conversations_df.index == conversation_id]['meta.donation_ee'].values[0]

    return 1 if er_donation + ee_donation > 0 else 0


### 3. Use the above functions to generate our raw features and labels

conversations = [get_conversation(c_id) for c_id in conversation_ids]
donations = [get_conversation_donation(c_id) for c_id in conversation_ids]

df_data = [[c, d] for (c, d) in zip(conversations, donations)]

### 4. Create a new dataframe out of our conversations

df = pd.DataFrame(df_data, columns=["conversation", "success"])

df.to_csv('pfg_df.csv', index=False)

## Data Cleaning

### 1. Convert utterances to lowercase

df['conversation'] = df['conversation'].str.lower()

### 2. Remove any HTML or URLs from the utterances

df['conversation'] = df['conversation'].apply(lambda x: re.sub('http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+', '', str(x)))
df['conversation'] = df['conversation'].apply(lambda x: re.sub('<[^<]+?>', '', str(x)))

### 3. Expand contractions

df['conversation'] = df['conversation'].apply(lambda x: contractions.fix(x))

### 4. Tokenization

from nltk.tokenize import sent_tokenize, word_tokenize
nltk.download('punkt', quiet=True)

### 4.1. Sentence tokenization

df['conversation'] = df['conversation'].apply(lambda x: ' '.join([w for w in sent_tokenize(x)]))

### 4.2. Word tokenization

df['conversation'] = df['conversation'].apply(lambda x: ' '.join([w for w in word_tokenize(x)]))

### 5. Remove non-alphanumeric characters

df['conversation'] = df['conversation'].apply(lambda x: re.sub('[^a-zA-Z0-9]', ' ', str(x)))

# Pre-processing

from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer

### 1. Remove the stop words

stop_words = stopwords.words('english')
df['conversation'] = df['conversation'].apply(lambda x: ' '.join([w for w in x.split() if w not in stop_words]))

### 2. Perform stemming

# stemmer = PorterStemmer()
# df['text'] = df['text'].apply(lambda x: ' '.join([stemmer.stem(w) for w in x.split()]))

### 3. Perform lemmatization

lemmatizer = nltk.WordNetLemmatizer()
df['conversation'] = df['conversation'].apply(lambda x: ' '.join([lemmatizer.lemmatize(w) for w in x.split()]))

### 4. Remove extra whitespace

df['conversation'] = df['conversation'].apply(lambda x: ' '.join(x.split()))

# Write the preprocessed dataframe to a .csv file

df.to_csv('pfg_df_cleaned.csv', index=False)
