==> success_classifier/4_bow_svm_classifier.ipynb <==
import numpy as np
import pandas as pd
from helper_funcs import report_f1_results, report_accuracy, split_data
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
import joblib

# Read in the dataframe(s)

df = pd.read_csv('cmv_df_cleaned.csv')
pfg_df = pd.read_csv('pfg_df_cleaned.csv')
our_df = pd.read_csv('our_df_cleaned.csv')

### Combine our CMV and PFG dataframes into a single dataframe

df = df.append(pfg_df, ignore_index=True)

# Filter out the rows that have NaN as a label

### Report the size of the cmv_df before and after filtering

size_before = len(df)
df = df[(df['success'] == 0) | (df['success'] == 1)]
size_after = len(df)
print(f"Size of the dataframe has been reduced from {size_before} down to {size_after}")

# Shuffle the dataframe

df = df.sample(frac=1).reset_index(drop=True)
df = df[:-1]

# Generate the labels array

labels = np.array(list(df['success']))

# Split the dataframe and labels array into smaller chunks

num_chunks = 8
dfs = np.split(df, num_chunks)
labels = np.split(labels, num_chunks)

# Bag of Words Feature Extraction

### 1. Create a list of unique words in all persuader conversations

unique_words = list(set([word for conversation in df["conversation"] for word in conversation.split()]))
unique_words_dict = {word: idx for idx, word in enumerate(unique_words)}

### 2. Use the unique words to create our BoW feature matrix

def generate_bow_feature(conversation):
    feature = [0] * len(unique_words)
    for word in conversation.split():
        idx = unique_words_dict[word] if word in unique_words_dict else None
        if idx is not None:
            feature[idx ] += 1

    return np.array(feature)

bow_matrices = []

for idx in range(len(dfs)):
    bow_matrices.append(np.array([generate_bow_feature(conversation) for conversation in dfs[idx]["conversation"]]))

# sklearn SVM Model

from sklearn.svm import SVC
from sklearn.model_selection import GridSearchCV

### 1. Write the training function

def train(features, labels):
    # parameters = [{'kernel': ['rbf'], 'C':[1, 10, 100, 1000], 'gamma': [0.001, 0.0001]}]

    svm = SVC(kernel='rbf', C=1000, gamma=0.001)
    # svm = SVC()
    # clf = GridSearchCV(svm, parameters, verbose=2)
    # clf.fit(features, labels)
    svm.fit(features, labels)

    # Return the trained model
    return svm

### 2. Write the testing function

def test(features, labels, model, feature_type="No Feature Type Provided"):
    label_predictions = model.predict(features)
    accuracy = accuracy_score(list(labels), label_predictions)
    report_accuracy(accuracy, feature_type, "sklearn SVM")

# Split data into testing and training subsets

train_xs = []
test_xs = []
train_ys = []
test_ys = []

for idx, matrix in enumerate(bow_matrices):
    train_x, test_x, \
        train_y, test_y = split_data(matrix, labels[idx])
    
    train_xs.append(train_x)
    test_xs.append(test_x)
    train_ys.append(train_y)
    test_ys.append(test_y)


# Testing the sklearn SVM Model

### 1. Train the model

for idx, (x, y) in enumerate(zip(train_xs, train_ys)):
    print(f"Training model {idx + 1} / {len(train_xs)}", end='\r')
    model = train(x, y)
    joblib.dump(model, f"./models/{idx}_bow_svm_model.pkl")
    del(model) # This might not be needed

# joblib.dump(model, "./models/0_tfidf_svm_model.pkl")
# joblib.load("./models/0_tfidf_svm_model.pkl")

### 2. Test the model

for idx, (x, y) in enumerate(zip(test_xs, test_ys)):
    model = joblib.load(f"./models/{idx}_bow_svm_model.pkl")
    test(test_xs[idx], test_ys[idx], model, f"Bag of Words Model #{idx + 1}")
    del(model) # This might not be needed

# Get predictions for our dataset

model = joblib.load(f"./models/1_bow_svm_model.pkl")

### 1. Generate the features for our conversations

# our_labels = [0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0]
# our_labels = [0, 0, 0, 0, 1, 0, 0, 1]
our_labels = [0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]

def get_accuracy(Y):
    total = len(our_labels)
    correct = 0
    for idx, label in enumerate(Y):
        if label == our_labels[idx]:
            correct = correct + 1
    
    return (correct / total) * 100


for i in range(1, 8, 1):
    model = joblib.load(f"./models/{i}_bow_svm_model.pkl")
    our_features = []
    for conversation in our_df["conversation"]:
        our_features.append(generate_bow_feature(conversation)[:-1])
    predictions = model.predict(our_features)
    print(f"[{get_accuracy(predictions):.2f}%] {predictions}")

for i in range(1, 8, 1):
    model = joblib.load(f"./models/{i}_bow_svm_model.pkl")
    our_features = []
    for conversation in our_df["conversation"]:
        our_features.append(generate_bow_feature(conversation)[:-1])
    predictions = model.predict(our_features)
    print(f"[{get_accuracy(predictions):.2f}%] {predictions}")

for i in range(1, 8, 1):
    model = joblib.load(f"./models/{i}_bow_svm_model.pkl")
    our_features = []
    for conversation in our_df["conversation"]:
        our_features.append(generate_bow_feature(conversation)[:-1])
    predictions = model.predict(our_features)
    print(f"[{get_accuracy(predictions):.2f}%] {predictions}")

our_predictions = model.predict(our_features)

our_predictions

our_df.tail()
